requirements.txt
# Data processing
pandas==2.1.3
numpy==1.26.2
pyspark==3.5.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# AWS Services
boto3==1.33.0
s3fs==2023.12.0

# Database
PyMySQL==1.1.0
sqlalchemy==2.0.23

# API
requests==2.31.0
python-dotenv==1.0.0

# Utilities
pytest==7.4.3
flake8==6.1.0
pytest-mock==3.12.0  
pyyaml==6.0.1


